requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
python-magic==0.4.27
google-cloud-storage==2.10.0
requests==2.31.0
elasticsearch==8.9.0
python-multipart==0.0.6
validators==0.22.0
gunicorn==21.2.0
supervisor==4.2.5
